refactor: moved rest package back to utils remove: Removed previous rate-limit code to make way for new code.
