+++ src/main/java/com/seailz/discordjar/action/sticker/ModifyStickerAction.java
@@ -3,7 +3,7 @@ package com.seailz.discordjar.action.sticker;
 import com.seailz.discordjar.DiscordJar;
 import com.seailz.discordjar.model.emoji.sticker.Sticker;
 import com.seailz.discordjar.utils.URLS;
-import com.seailz.discordjar.rest.DiscordRequest;
+import com.seailz.discordjar.utils.rest.DiscordRequest;
 import org.json.JSONObject;
 import org.springframework.web.bind.annotation.RequestMethod;
