+++ src/main/java/com/seailz/discordjar/events/model/interaction/command/MessageContextCommandInteractionEvent.java
@@ -3,8 +3,8 @@ package com.seailz.discordjar.events.model.interaction.command;
 import com.seailz.discordjar.DiscordJar;
 import com.seailz.discordjar.model.message.Message;
 import com.seailz.discordjar.utils.URLS;
-import com.seailz.discordjar.rest.DiscordRequest;
-import com.seailz.discordjar.rest.DiscordResponse;
+import com.seailz.discordjar.utils.rest.DiscordRequest;
+import com.seailz.discordjar.utils.rest.DiscordResponse;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.json.JSONObject;
