+++ src/main/java/com/seailz/discordjar/model/application/Application.java
@@ -7,8 +7,8 @@ import com.seailz.discordjar.model.team.Team;
 import com.seailz.discordjar.model.user.User;
 import com.seailz.discordjar.utils.Checker;
 import com.seailz.discordjar.utils.URLS;
-import com.seailz.discordjar.rest.DiscordRequest;
-import com.seailz.discordjar.rest.DiscordResponse;
+import com.seailz.discordjar.utils.rest.DiscordRequest;
+import com.seailz.discordjar.utils.rest.DiscordResponse;
 import com.seailz.discordjar.utils.Snowflake;
 import com.seailz.discordjar.utils.flag.BitwiseUtil;
 import org.jetbrains.annotations.NotNull;
