+++ src/main/java/com/seailz/discordjar/model/interaction/callback/internal/InteractionHandlerImpl.java
@@ -6,7 +6,7 @@ import com.seailz.discordjar.action.interaction.followup.InteractionFollowupActi
 import com.seailz.discordjar.model.interaction.callback.InteractionHandler;
 import com.seailz.discordjar.model.message.Message;
 import com.seailz.discordjar.utils.URLS;
-import com.seailz.discordjar.rest.DiscordRequest;
+import com.seailz.discordjar.utils.rest.DiscordRequest;
 import org.json.JSONObject;
 import org.springframework.web.bind.annotation.RequestMethod;
