+++ src/main/java/com/seailz/discordjar/model/user/User.java
@@ -8,8 +8,8 @@ import com.seailz.discordjar.utils.CDNAble;
 import com.seailz.discordjar.utils.Mentionable;
 import com.seailz.discordjar.utils.StringFormatter;
 import com.seailz.discordjar.utils.URLS;
-import com.seailz.discordjar.rest.DiscordRequest;
-import com.seailz.discordjar.rest.DiscordResponse;
+import com.seailz.discordjar.utils.rest.DiscordRequest;
+import com.seailz.discordjar.utils.rest.DiscordResponse;
 import com.seailz.discordjar.utils.flag.BitwiseUtil;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
