+++ src/main/java/com/seailz/discordjar/oauth2/LinkedRoles.java
@@ -3,8 +3,8 @@ package com.seailz.discordjar.oauth2;
 import com.seailz.databaseapi.Database;
 import com.seailz.discordjar.DiscordJar;
 import com.seailz.discordjar.utils.URLS;
-import com.seailz.discordjar.rest.DiscordRequest;
-import com.seailz.discordjar.rest.DiscordResponse;
+import com.seailz.discordjar.utils.rest.DiscordRequest;
+import com.seailz.discordjar.utils.rest.DiscordResponse;
 import org.json.JSONObject;
 import org.slf4j.bridge.SLF4JBridgeHandler;
 import org.springframework.boot.SpringApplication;
