--- src/main/java/com/seailz/discordjar/rest/RateLimit.java
@@ -1,8 +1,0 @@
-package com.seailz.discordjar.rest;
-
-public record RateLimit(
-        double limit,
-        double remaining,
-        double resetAfter
-) {
-}
