+++ src/main/java/com/seailz/discordjar/utils/rest/DiscordResponse.java
@@ -1,4 +1,4 @@
-package com.seailz.discordjar.rest;
+package com.seailz.discordjar.utils.rest;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
