+++ src/main/java/com/seailz/discordjar/utils/rest/RequestQueueHandler.java
@@ -1,0 +1,20 @@
+package com.seailz.discordjar.utils.rest;
+
+import com.seailz.discordjar.DiscordJar;
+
+/**
+ * A class that handles the queue of requests to the Discord API.
+ *
+ * @author Seailz
+ * @see com.seailz.discordjar.utils.rest.DiscordRequest
+ * @since 1.0
+ */
+public class RequestQueueHandler {
+
+    public RequestQueueHandler(DiscordJar jv) {
+        new Thread(() -> {
+
+        }).start();
+    }
+
+}
